requirements.airflow-2.0.2.txt
apache-airflow-providers-google==3.0.0
facebook-business==13.0.0
mysql-connector-python==8.0.24
paramiko==2.8.1
pg8000==1.26.0
